Implement delete button

+++ public/js/saved.js
@@ -1,11 +1,15 @@
-const remove = document.querySelector('.footer');
+const remove = document.querySelector(".saved-container");
 
-remove.addEventListener('click', (e) => {
-    const endpoint = '/api/user/saved-movies';
-    fetch(endpoint, {
-        method: 'DELETE'
-    })
-    .then((response) => response.json()) 
-    .then((data => console.log(data)))
-    .catch(err => console.log(err));
+remove.addEventListener("click", async (e) => {
+  if (!e.target.matches(".footer button")) return;
+  const endpoint = "/api/user/saved-movies/" + e.target.dataset.movie;
+  const response = await fetch(endpoint, {
+    method: "DELETE",
+  });
+  if (response.ok) {
+    document.location.reload();
+  } else {
+    const data = await response.json();
+    renderErrors(data.errors);
+  }
 });
